pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flipswitch-sdk"
version = "0.1.0"
description = "Flipswitch SDK with real-time SSE support for OpenFeature"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
keywords = ["feature-flags", "openfeature", "flipswitch", "sse", "real-time"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "openfeature-sdk>=0.7.0",
    "openfeature-provider-ofrep>=0.2.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-httpserver>=1.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.urls]
Homepage = "https://github.com/flipswitch-dev/flipswitch"
Documentation = "https://flipswitch.dev/docs/sdks/python"
Repository = "https://github.com/flipswitch-dev/flipswitch"

[tool.setuptools.packages.find]
where = ["."]
include = ["flipswitch*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
